requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==2.3.0
Django==5.0
djangorestframework==3.15.2
factory_boy==3.3.3
Faker==37.1.0
idna==3.10
psycopg-binary==3.2.7
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
